+++ capi/autd3capi-link-twincat/src/lib.rs
@@ -4,7 +4,7 @@
  * Created Date: 27/05/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 10/11/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -18,7 +18,7 @@ use std::{
     time::Duration,
 };
 
-use autd3capi_def::{common::*, LinkBuilderPtr};
+use autd3capi_def::*;
 
 use autd3_link_twincat::{local::twincat_link::*, remote::remote_twincat_link::*};
 
@@ -92,7 +92,7 @@ pub unsafe extern "C" fn AUTDLinkRemoteTwinCAT(
         Err(e) => {
             let err = e.to_string();
             ResultLinkRemoteTwinCATBuilder {
-                result: LinkRemoteTwinCATBuilderPtr(NULL),
+                result: LinkRemoteTwinCATBuilderPtr(std::ptr::null()),
                 err_len: err.as_bytes().len() as u32 + 1,
                 err: Box::into_raw(Box::new(err)) as _,
             }