+++ capi/autd3capi/src/gain/custom.rs
@@ -4,14 +4,14 @@
  * Created Date: 23/08/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 23/11/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
  *
  */
 
-use autd3capi_def::{common::*, take_gain, Drive, GainPtr};
+use autd3capi_def::*;
 
 #[no_mangle]
 #[must_use]
@@ -29,7 +29,7 @@ pub unsafe extern "C" fn AUTDGainCustomSet(
     ptr: *const Drive,
     len: u32,
 ) -> GainPtr {
-    let mut drives = Vec::<autd3capi_def::common::driver::common::Drive>::with_capacity(len as _);
+    let mut drives = Vec::<autd3capi_def::driver::common::Drive>::with_capacity(len as _);
     drives.set_len(len as _);
     std::ptr::copy_nonoverlapping(ptr as *const _, drives.as_mut_ptr(), len as _);
     GainPtr::new(take_gain!(custom, CustomGain).set(dev_idx as _, drives))