+++ capi/autd3capi/src/gain/uniform.rs
@@ -4,17 +4,14 @@
  * Created Date: 24/08/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 23/11/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
  *
  */
 
-use autd3capi_def::{
-    common::{autd3::gain::Uniform, *},
-    take_gain, GainPtr,
-};
+use autd3capi_def::{autd3::gain::Uniform, *};
 
 #[no_mangle]
 #[must_use]