+++ capi/autd3capi/src/link/mod.rs
@@ -4,17 +4,14 @@
  * Created Date: 24/08/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 25/10/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
  *
  */
 
-use autd3capi_def::{
-    common::{cast, Cnt},
-    ControllerPtr, LinkPtr,
-};
+use autd3capi_def::{cast, Cnt, ControllerPtr, LinkPtr};
 
 pub mod audit;
 pub mod nop;