+++ dotnet/cs/src/Link/Audit.cs
@@ -4,7 +4,7 @@
  * Created Date: 22/09/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 24/11/2023
+ * Last Modified: 29/11/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -12,6 +12,7 @@
  */
 
 using AUTD3Sharp.Internal;
+using AUTD3Sharp.NativeMethods;
 using System;
 
 #if UNITY_2020_2_OR_NEWER