+++ python/pyautd3/modulation/sine.py
@@ -11,7 +11,7 @@ Copyright (c) 2023 Shun Suzuki. All rights reserved.
 
 """
 
-
+from pyautd3.emit_intensity import EmitIntensity
 from pyautd3.internal.modulation import IModulationWithSamplingConfig
 from pyautd3.native_methods.autd3capi import NativeMethods as Base
 from pyautd3.native_methods.autd3capi_def import ModulationPtr
@@ -21,15 +21,15 @@ class Sine(IModulationWithSamplingConfig):
     """Sine wave modulation."""
 
     _freq: int
-    _amp: float | None
-    _offset: float | None
+    _intensity: EmitIntensity | None
+    _offset: EmitIntensity | None
     _phase: float | None
 
     def __init__(self: "Sine", freq: int) -> None:
         """Constructor.
 
         The sine wave is defined as `amp/2 * sin(2π * freq * t + phase) + offset`,
-        where `t` is time, and `amp = 1`, `phase = 0`, `offset = 0.5` by default.
+        where `t` is time, and `amp = EmitIntensity.maximum()`, `phase = 0`, `offset = EmitIntensity.maximum()/2` by default.
 
         Arguments:
         ---------
@@ -37,28 +37,28 @@ class Sine(IModulationWithSamplingConfig):
         """
         super().__init__()
         self._freq = freq
-        self._amp = None
+        self._intensity = None
         self._offset = None
         self._phase = None
 
-    def with_amp(self: "Sine", amp: float) -> "Sine":
-        """Set amplitude.
+    def with_intensity(self: "Sine", intensity: int | EmitIntensity) -> "Sine":
+        """Set intensity.
 
         Arguments:
         ---------
-            amp: Amplitude
+            intensity: Intensity
         """
-        self._amp = amp
+        self._intensity = EmitIntensity._cast(intensity)
         return self
 
-    def with_offset(self: "Sine", offset: float) -> "Sine":
+    def with_offset(self: "Sine", offset: int | EmitIntensity) -> "Sine":
         """Set offset.
 
         Arguments:
         ---------
             offset: Offset
         """
-        self._offset = offset
+        self._offset = EmitIntensity._cast(offset)
         return self
 
     def with_phase(self: "Sine", phase: float) -> "Sine":
@@ -73,10 +73,10 @@ class Sine(IModulationWithSamplingConfig):
 
     def _modulation_ptr(self: "Sine") -> ModulationPtr:
         ptr = Base().modulation_sine(self._freq)
-        if self._amp is not None:
-            ptr = Base().modulation_sine_with_amp(ptr, self._amp)
+        if self._intensity is not None:
+            ptr = Base().modulation_sine_with_intensity(ptr, self._intensity.value)
         if self._offset is not None:
-            ptr = Base().modulation_sine_with_offset(ptr, self._offset)
+            ptr = Base().modulation_sine_with_offset(ptr, self._offset.value)
         if self._phase is not None:
             ptr = Base().modulation_sine_with_phase(ptr, self._phase)
         if self._config is not None: